moved validate_transaction
